Resource Functions and Adding to Rocket
Adjusted the functions that the resources use to not take any parameters
and to return void.
Changed how the time is checked for those functions to private floats
instead of floats within the function directly.
Added more comments to the variables and functions of the resources.

Adjusted how the function to add a part to the Rocket works. Now checks
if the part is in the list first before removing or adding the part. Not
yet fully implemented.