Work on harvester
Harvester can now pick up food and drop it off at the silo.
Decontamination code has not been implemented yet. The current state of
the harvester is that it can now harvest food, bring it to the silo and
go back to the recent clicked tree to continue collecting or it will
wait in idle for instructions if the tree has been destroyed or if the
tree is out of resources.