More Rocket Part Work
Adjusted the interface for the rocket parts to use unsigned integers.
Changed the related properties and variables int the part classes.
Once more adjusted how each part is added to the list. Now implementing
a secondary list in each class that's checked before an object is
removed. No longer throws an error related to when a part is removed
from the list.
Currently takes input to increment the users Steel and Fuel counts.