Added better pickup functionality
The miner and extractor have better pickup functionality. They stock
resources accordingly. If the user stops a unit currently stocking then
goes back later to continue the stocking, it will pickup where it left
off and the count will be correct. Next will be giving this funtionality
to the harvester.