Update enemy state behavior
Enemies now search when in the idle state for the closest unit or
resource that can be tainted. Spawnpoints do not respawn enemies until a
unit walks into the trigger after 15 seconds has passed. Updated the
camera controller so there is a boolean allowing the user to pan the
camera with mouse position as well.