Got Caught Up
I had a conflict in a pull request that turned into much more of a
problem than I realized.
But all is fixed now and up to date. I updated the decontamination
prefab to have an appropriate sized box collider and a front door child
object for the units to go to and decontaminate resources.