Fix Camera Commit
Camera can now be panned with the mouse by toggling that functionality
with pressing the y button on the keyboard. The camera rotation is
currently locked at a 45 degree rotation to the left or right. Project
settings only build to aspect ratio of 16:9. Took off box collider of
decontamination tent and added mesh collider. Changed default size of
cameras orthographic size to 6.