Fixed some bugs and made a test scene
Fixed bugs that were preventing proper button click functionality on the
UI while interacting with units using the mouse clicks. Added a test
scene for me to experiment with the UI some more. Im currently working
on a piece of UI that will be populated based on the units selected.
Further updates are needed. But base functionality has been started.