Working on some UI
Working on making a UI panel that populates with buttons for each unit
the user has selected. Also, working on implementing a unit stat upgrade
system. Took out code that I wasnt using. Fixed bugs that were causing
the units be selected when trying to click on a UI Element.