The Upgrade Panel and Tooltip Commit
Added functionality of a tooltip panel. When hovering over the populated
unit button, the tooltip panel pops up, showing the user the current
stats of that specific unit. When not hovering anymore over the button,
then the tooltip panel turns off. Also added functionality of the user
being able to now click on the button that is populated for a specific
unit. When clicked on, an upgrade panel will pop open. The units
information will populate the opened panel. Upcoming updates will allow
the user to upgrade specific stats on the screen for the given unit.
Cleaned up some code and added the functionality for the user to obtain
reward points when completing an objective. This will undergo further
updates to determine exactly how many points to obtain for the completed
objective.