Stat Upgrade Commit 1
Added functionality for a unit to receive a stat increase based on the
choice of the user on the upgrade panel. This will need to undergo
further changes. It is poorly optimized at the moment but I will fix it
in upcoming commits. Also, Added the functionality of the Call home
button. This button will find all units on the map and send them back to
the barracks.