Code Refactor
Went through all the scripts I created and took out code that was not
being used and made the appropriate fields private that should have
been. Added comments to the ones that did not have any.  Created a stats
class that will be used to reference a units stats. Updated the rotation
of the units to now be able to look at there move destination
accordingly. Ran test cases on upgrading a units stats in the upgrade
panel. The functionality for stat upgrades is now fully functional.
Fixed a big issue with the nav mesh baking that was causing tree nodes
to be baked with having their limbs being referenced in the bake. This
was causing crowding issues with the units.