Enemy AI Update
The prior useage of Enemy AI was very expensive and wasnt working
properly. Started a new implementation today. Current state is the
aliens will follow a target while the target is in the radius of the
enemy. When a target gets out of range the enemy will stop pursuing.
This will undergo further updates. Still needs implementation of
tainting a resource.