Added Unit healing
A unit can now be healed. By bringing a cooked food object over the unit
and right clicking the unit, it will heal the unit only if its health is
lower than its max health. Made a Hover script and attached it to the
cooked food prefab. this will deactivate the cooked food prefabs
collider so the mouse can still raycast through it and this script
handles the movement of the prefab setting its position to the mouse
position.