UI Implementation and Changing Code
Setup Dante's UI in the scene I work with.
Worked on adjusting the code for adding and removing parts to make it
more suitable for being executed by an OnClick() event.
Made a new function for adding and replacing a specific part to the
list. It seems that having individual functions for this might be the
most efficient.