Adding Parts Overhaul
Overhauled the function for adding parts to the list and replacing
current ones. New function now works well with multiple objects of the
same type.

All the input commands in the update are for testing purposes. Should be
good for implementation into the workshop buttons, if they're setup to
instantiate an object of an acceptible type. Though the price checks
might need to be moved to the functions for the buttons.