Setup hotkey functionality
The hotkeys on the customize menu that will be customizable, are now
mapped to the appropriate key. They have been tested and work
accrodingly.
The customize menu is subject to change over the course of the project.
Their will be removal of some keys from the menu and some will be added.
Updated the GameUI prefab to now have appropriate actions for the
corresponding buttons the user clicked on.
The functionality for changing a hotkey has not been implemented yet.