Added arrow key functionality to camera movement and more comment fixes
Fixed some comments throughout different scripts.
Cleaned up some code in a few scripts. Added functionality in the Camera
Controller script that will allow the user to move the camera with the
arrow keys. Made the move speed variable for the camera public to
provide later functionality of the camera speed slider in the options
menu to become updated accordingly. Also, allowing the Move Speed to be
updated when the slider is moved.