Updated GameUI prefab
Fixed the main menu scene to allow hotkey change functionality, pre
baked the lighting in the main game scene to fix the lighting issue upon
being loaded. Added functionality to the key bind script to allow
carrying over of the new hot key changes if the user changes them in the
main menu.