Building Cockpits and Wings
Added a string to the IRocketParts interface to give the parts names.
Removed the Chassis and BaseChassis classes due to no longer needing
them.
Wrote functions for building and replacing the cockpit and wings. Made
additional functions to make use of the functions simpler in other
classes.
Changed the AddPart function to a bool rather than a void for checks
before it's formally called. Implemented additional checks to check if a
list contains objects of the same name as the parameter.