Setup functionality for spawning units
Setup functionality for spawning units. Added the new miner into the
scene and textured them all correctly. When hovering over the buttons in
the Units panel, the highlight will now light up the color according to
the unit it represents. Setup functionality for further updates of
changing the orb color of a unit based on health percentage,