Updated tooltip functionality
Added a building script that is attached to buildings in game. When
hovering over the building, it will update the tool tip panel with its
description. Updated the tool tip panel to now be able to be
disabled/enabled through the options menu. Fixed a bug that was causing
units to have the selection square beneath to be active when they are
spawned.